move SpongeFunction creators to curl and kerl package
