+++ signing/legacy/signing_test.go
@@ -2,8 +2,8 @@ package signing_test
 
 import (
 	. "github.com/iotaledger/iota.go/consts"
+	. "github.com/iotaledger/iota.go/curl"
 	. "github.com/iotaledger/iota.go/signing/legacy"
-	. "github.com/iotaledger/iota.go/signing/utils"
 	. "github.com/iotaledger/iota.go/trinary"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
