+++ signing/signing_milestone_test.go
@@ -2,9 +2,9 @@ package signing_test
 
 import (
 	"github.com/iotaledger/iota.go/consts"
+	. "github.com/iotaledger/iota.go/curl"
 	"github.com/iotaledger/iota.go/merkle"
 	. "github.com/iotaledger/iota.go/signing"
-	. "github.com/iotaledger/iota.go/signing/utils"
 	. "github.com/iotaledger/iota.go/transaction"
 	. "github.com/iotaledger/iota.go/trinary"
 	. "github.com/onsi/ginkgo"
