+++ signing/utils/sponge.go
@@ -1,10 +1,7 @@
-// Package signing provides functions for creating and validating essential cryptographic
-// components in IOTA, such as subseeds, keys, digests and signatures.
+// Package sponge provides an interface for the sponge functions in IOTA.
 package sponge
 
 import (
-	"github.com/iotaledger/iota.go/curl"
-	"github.com/iotaledger/iota.go/kerl"
 	. "github.com/iotaledger/iota.go/trinary"
 )
 
@@ -17,22 +14,7 @@ type SpongeFunction interface {
 	Reset()
 }
 
-// NewCurlP27 returns a new CurlP27.
-func NewCurlP27() SpongeFunction {
-	return curl.NewCurl(curl.CurlP27)
-}
-
-// NewCurlP81 returns a new CurlP81.
-func NewCurlP81() SpongeFunction {
-	return curl.NewCurl(curl.CurlP81)
-}
-
-// NewKerl returns a new Kerl.
-func NewKerl() SpongeFunction {
-	return kerl.NewKerl()
-}
-
-// GetSpongeFunc checks if a hash function was given, otherwise uses defaultSpongeFuncCreator, or Kerl.
+// GetSpongeFunc checks if a hash function was given, otherwise uses defaultSpongeFuncCreator. Panics if none given.
 func GetSpongeFunc(spongeFunc []SpongeFunction, defaultSpongeFuncCreator ...SpongeFunctionCreator) SpongeFunction {
 	if len(spongeFunc) > 0 {
 		return spongeFunc[0]
@@ -40,5 +22,5 @@ func GetSpongeFunc(spongeFunc []SpongeFunction, defaultSpongeFuncCreator ...Spon
 	if len(defaultSpongeFuncCreator) > 0 {
 		return defaultSpongeFuncCreator[0]()
 	}
-	return NewKerl()
+	panic("No sponge function given")
 }
